updated deap_genetic_guide, need more testing
I'm back to the configuration of the "lists of lists", now I have to check that the internal functions of DGG are working properly.
The past functions of "generate" and "evaluate" work. I used DGG instead of DGGGP because I couldn't think of a method to use the latter, after all DGGGP is a representation of a certain problem (the shortest route from one city to another).